Calculate Monthly DHW Energy Demand per Square Meter

Q_DHW_Demand estimates the monthly energy required for Domestic Hot Water (DHW) per square meter of the building. The function calculates this demand based on the daily DHW demand specific to the building type.

Usage

Q_DHW_Demand(params)

Arguments

params

A list returned by the get_parameters function.

Value

A numeric vector of length 12, representing the monthly energy demand for DHW per square meter, measured in megajoules (MJ). Each element of the vector corresponds to one month of the year.