Calculate CO2 emissions

calculate_emissions calculates the CO2 emissions of a building according to the input arguments described below.

Usage

calculate_emissions(
  area,
  floors,
  year,
  utilisation_key,
  climate_code,
  energy_carrier,
  walls_refurb_year = NULL,
  roof_refurb_year = NULL,
  windows_refurb_year = NULL,
  floor_refurb_year = NULL,
  heating_install_year = NULL
)

Arguments

area

Mandatory, numeric. Area (in square meters) where heating or cooling is required ("Energiebezugsfläche").

e.g. 1000

floors

Mandatory, numeric. Number of floors within the building envelope ("thermische Gebäudehülle").

e.g. 4

year

Mandatory, numeric. Year of building construction.

e.g. 1981

utilisation_key

Mandatory, numeric. Main utilisation of the building according to SIA 380/1:2009.

The following utilisations are possible:

ID	Name	Description

1	Wohnen Mehrfamilienhaus	Mehrfamilienhäuser, Alterssiedlungen und -wohnungen, Hotels, Mehrfamilien-Ferienhäuser und Ferienheime, Kinder- und Jugendheime, Tagesheime, Behindertenheime, Drogenstationen, Kasernen, Strafanstalten

2	Wohnen Einfamilienhaus	Ein- und Zweifamilienhäuser, Ein- und ZweifamilienFerienhäuser, Reiheneinfamilienhäuser

3	Verwaltung	private und öffentliche Bürobauten, Schalterhallen, Arztpraxen, Bibliotheken, Ateliers, Ausstellungsbauten, Kulturzentren, Rechenzentren, Fernmeldegebäude, Fernsehgebäude, Filmstudios

4	Schulen	Gebäude für Schulen aller Stufen, Kindergärten und - horte, Schulungsräume, Ausbildungszentren, Kongressgebäude, Labors, Forschungsinstitute, Gemeinschaftsräume, Freizeitanlagen

5	Verkauf	Verkaufsräume aller Art inkl. Einkaufszentren, Messegebäude

6	Restaurants	Restaurants (inkl. Küchen), Cafeterias, Kantinen, Dancings, Diskotheken

7	Versammlungslokale	Theater, Konzertsäle, Kinos, Kirchen, Abdankungshallen, Aulas, Sporthallen mit viel Publikum

8	Spitäler	Spitäler, psychiatrische Kliniken, Krankenheime, Altersheime, Rehabilitationszentren, Behandlungsräume

9	Industrie	Fabrikationsgebäude, Gewerbebauten, Werkstätten, Servicestationen, Werkhöfe, Bahnhöfe, Feuerwehrgebäude

10	Lager	Lagerhallen, Verteilzentren

11	Sportbauten	Turn- und Sporthallen, Gymnastikräume, Tennishallen, Kegelbahnen, Fitnesszentren, Sportgarderoben

12	Hallenbäder	Hallenbäder, Lehrschwimmbecken, Saunagebäude, Heilbäder

Check the full dataset using utilisation.

climate_code

Mandatory, string. Abbreviation code of the assigned climate station according to "Auszug aus Merkblatt SIA 2028:2015". The following climate stations are available:

ABO: Adelboden

AIG: Aigle

ALT: Altdorf

BAS: Basel-Binningen

BER: Bern-Liebefeld

BUS: Buchs-Aarau

CDF: La Chaux-de-Fonds

CHU: Chur

DAV: Davos

DIS: Disentis

ENG: Engelberg

FRE: La Frétaz

GLA: Glarus

GSB: Grand-St-Bernard

GUT: Güttingen

GVE: Genève-Cointrin

INT: Interlaken

KLO: Zürich-Kloten

LUG: Lugano

LUZ: Luzern

MAG: Magadino

MVE: Montana

NEU: Neuchâtel

OTL: Locarno-Monti

PAY: Payerne

PIO: Piotta

PUY: Pully

ROB: Robbia

RUE: Rünenberg

SAM: Samedan

SBE: San Bernardino

SCU: Scuol

SHA: Schaffhausen

SIO: Sion

SMA: Zürich-MeteoSchweiz

STG: St. Gallen

ULR: Ulrichen

VAD: Vaduz

WYN: Wynau

ZER: Zermatt

Check the full dataset using climate.

energy_carrier

Mandatory, string. Energy carrier used for heating and domestic hot water production.

The following energy carriers are available:

oilHeating

gasHeating

other

walls_refurb_year

Optional, numeric, default value NULL. Year of wall refurbishment.

e.g. 2010

Supply NULL or NA if no refurbishment has taken place since the year of building construction.

roof_refurb_year

Optional, numeric, default value NULL. Year of roof refurbishment.

e.g. 2010

Supply NULL or NA if no refurbishment has taken place since the year of building construction.

windows_refurb_year

Optional, numeric, default value NULL. Year of window refurbishment.

e.g. 2010

Supply NULL or NA if no refurbishment has taken place since the year of building construction.

floor_refurb_year

Optional, numeric, default value NULL. Year of basement floor refurbishment.

e.g. 2010

Supply NULL or NA if no refurbishment has taken place since the year of building construction.

heating_install_year

Optional, numeric, default value NULL. Year of heating installation.

e.g. 2010

Supply NULL or NA if no heating replacement has taken place since the year of building construction.

Value

calculate_emissions returns a list r containing the calculation results:

r$heatEnergy: annual heat energy required per area, in MJ/m2 per year

r$emissionCoefficient: applied CO2 emission coefficient according to the given energy carrier, in kg/MJ

r$emissionsPerArea: annual CO2 emissions per area, in kg/m2 per year

r$emissionsTotal: total annual CO2 emissions, in kg per year

Examples

# Example 1:
# calculate CO2 emissions of a multi-family house built in 1981, without refurbishments
ex1 <- calculate_emissions(
  area = 1000,
  floors = 4,
  year = 1981,
  utilisation_key = 1,
  climate_code = "KLO",
  energy_carrier = "gasHeating"
)
str(ex1)
#> List of 4
#>  $ heatEnergy         : num 458
#>  $ emissionCoefficient: num 0.064
#>  $ emissionsPerArea   : num 29.3
#>  $ emissionsTotal     : num 29343

# Example 2:
# same building as above but with additional refurbishments (roof & walls) in 2010
ex2 <- calculate_emissions(
  area = 1000,
  floors = 4,
  year = 1981,
  utilisation_key = 1,
  climate_code = "KLO",
  energy_carrier = "gasHeating",
  walls_refurb_year = 2010,
  roof_refurb_year = 2010,
  heating_install_year = 2010
)
str(ex2)
#> List of 4
#>  $ heatEnergy         : num 302
#>  $ emissionCoefficient: num 0.0599
#>  $ emissionsPerArea   : num 18.1
#>  $ emissionsTotal     : num 18107