Gather and Validate Parameters for CO2 Emission Calculations

This function validates the input parameters and prepare the input_args list for further processing . by the getParams function. It ensures that all mandatory parameters are present and correctly formatted, and it collects information about any refurbishments that have been carried out.

get_parameters(
  area,
  floors,
  year,
  utilisation_key,
  climate_code,
  energy_carrier,
  walls_refurb_year = NULL,
  roof_refurb_year = NULL,
  windows_refurb_year = NULL,
  floor_refurb_year = NULL,
  heating_install_year = NULL
)

Arguments

area

Mandatory, numeric. Area (in square meters) where heating or cooling is required ("Energiebezugsfläche").

e.g. 1000

floors

Mandatory, numeric. Number of floors within the building envelope ("thermische Gebäudehülle").

e.g. 4

year

Mandatory, numeric. Year of building construction.

e.g. 1981

utilisation_key

Mandatory, numeric. Main utilisation of the building according to SIA 380/1:2009.

The following utilisations are possible:

1: Wohnen Mehrfamilienhaus

2: Wohnen Einfamilienhaus

3: Büro

4: Schulen

5: Verkauf

6: Restaurants

7: Versammlungslokale

8: Spitäler

9: Industrie

10: Lager

11: Sportbauten

12: Hallenbäder

Check the full dataset using utilisation.

climate_code

Mandatory, string. Abbreviation code of the assigned climate station according to "Auszug aus Merkblatt SIA 2028:2015".

The following climate stations are available:

ABO: Adelboden

AIG: Aigle

ALT: Altdorf

BAS: Basel-Binningen

BER: Bern-Liebefeld

BUS: Buchs-Aarau

CDF: La Chaux-de-Fonds

CHU: Chur

DAV: Davos

DIS: Disentis

ENG: Engelberg

FRE: La Frétaz

GLA: Glarus

GSB: Grand-St-Bernard

GUT: Güttingen

GVE: Genève-Cointrin

INT: Interlaken

KLO: Zürich-Kloten

LUG: Lugano

LUZ: Luzern

MAG: Magadino

MVE: Montana

NEU: Neuchâtel

OTL: Locarno-Monti

PAY: Payerne

PIO: Piotta

PUY: Pully

ROB: Robbia

RUE: Rünenberg

SAM: Samedan

SBE: San Bernardino

SCU: Scuol

SHA: Schaffhausen

SIO: Sion

SMA: Zürich-MeteoSchweiz

STG: St. Gallen

ULR: Ulrichen

VAD: Vaduz

WYN: Wynau

ZER: Zermatt

Check the full dataset using climate.

energy_carrier

Mandatory, string. Energy carrier used for heating and domestic hot water production.

The following energy carriers are available:

oilHeating

gasHeating

other

walls_refurb_year

Optional, numeric, default value NULL. Year of wall refurbishment.

e.g. 2010

Supply NULL or NA if no refurbishment has taken place since the year of building construction.

roof_refurb_year

Optional, numeric, default value NULL. Year of roof refurbishment.

e.g. 2010

Supply NULL or NA if no refurbishment has taken place since the year of building construction.

windows_refurb_year

Optional, numeric, default value NULL. Year of window refurbishment.

e.g. 2010

Supply NULL or NA if no refurbishment has taken place since the year of building construction.

floor_refurb_year

Optional, numeric, default value NULL. Year of basement floor refurbishment.

e.g. 2010

Supply NULL or NA if no refurbishment has taken place since the year of building construction.

heating_install_year

Optional, numeric, default value NULL. Year of heating installation.

e.g. 2010

Supply NULL or NA if no heating replacement has taken place since the year of building construction.

Value

A list containing all validated and formatted input parameters, including any refurbishment details.

Examples

params <- get_parameters(
  area = 1000,
  floors = 5,
  year = 1980,
  utilisation_key = 3,
  climate_code = "ZER",
  energy_carrier = "gasHeating",
  walls_refurb_year = 2010,
  roof_refurb_year = 2015
)