Calculate Total Monthly Heat Losses

The function applies formula 99 of the SIA 380/1 norm, and aligns with varaible Q_ot.
Q_Losses_Total calculates the total monthly heat losses per square meter of the building. This function combines heat losses due to transmission (Q_Losses_Transmission) and ventilation (Q_Losses_Ventilation), providing a comprehensive view of the building's overall heat loss. The calculation reflects the sum of these two components, giving an estimate of the total energy loss through the building envelope and due to air exchange.

Usage

Q_Losses_Total(params)

Arguments

params

A list returned by the get_parameters function.

Value

A numeric vector of length 12, representing the total heat losses for each month, measured in megajoules (MJ) per square meter. Each element corresponds to one month of the year, illustrating the seasonal fluctuation in heat losses.